+++ Cliente1/Cliente1/App_Start/WebApiConfig.cs
@@ -9,6 +9,9 @@ namespace Cliente1
     {
         public static void Register(HttpConfiguration config)
         {
+            // Configuración y servicios de API web
+
+            // Rutas de API web
             config.MapHttpAttributeRoutes();
 
             config.Routes.MapHttpRoute(
