+++ Cliente1/Cliente1/Controllers/HomeController.cs
@@ -1,5 +1,4 @@
-﻿using Cliente1.Models;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
@@ -9,8 +8,6 @@ namespace Cliente1.Controllers
 {
     public class HomeController : Controller
     {
-
-
         public ActionResult Index()
         {
             return View();
@@ -29,6 +26,5 @@ namespace Cliente1.Controllers
 
             return View();
         }
-
     }
 }
