+++ Cliente1/Cliente1/Models/ProductoData.cs
@@ -8,7 +8,7 @@ using System.Web;
 
 namespace Cliente1.Models
 {
-    public class PriductoData
+    public class ProductoData
     {
         private SqlConnection _con;
 
@@ -53,6 +53,37 @@ namespace Cliente1.Models
             using (SqlCommand command = new SqlCommand("SelectProduct", _con))
             {
                 command.CommandType = CommandType.StoredProcedure;
+                _con.Open();
+                SqlDataReader reader = command.ExecuteReader();
+
+                while (reader.Read())
+                {
+                    Producto p = new Producto()
+                    {
+                        id_produto = Convert.ToInt32(reader["id_producto"]),
+                        nombre = Convert.ToString(reader["nombre"]),
+                        precio = Convert.ToInt32(reader["precio"]),
+                        url_img = Convert.ToString(reader["url_img"]),
+                        cantidad = Convert.ToInt32(reader["cantidad"]),
+                        estado = Convert.ToByte(reader["estado"]),
+                        descripcion = Convert.ToString(reader["descripcion"]),
+                    };
+                    productolista.Add(p);
+                }
+                _con.Close();
+                return productolista;
+            }
+        }
+
+        public List<Producto> ObtenerProductoId(int id_produto)
+        {
+            Connection();
+            List<Producto> productolista = new List<Producto>();
+
+            using (SqlCommand command = new SqlCommand("SelectProductoID", _con))
+            {
+                command.CommandType = CommandType.StoredProcedure;
+                command.Parameters.AddWithValue("@id", id_produto);
                 _con.Open();
                 SqlDataReader reader = command.ExecuteReader();
